// frontend/types.ts
// User Registration Information
export type UserRegistration = {
    clerk_id: string
    first_name: string
    last_name: string
    email: string
    school: string
    state: string
    grade_level: string
    current_score: number
    dream_score: number
    test_date: string
    subscription: string
    proficiency: string
    referral: string
}